Move authenticate_user to newflow module
